+++ app/src/main/java/application_projet4_groupe12/entities/ShopFrame.java
@@ -2,21 +2,21 @@ package application_projet4_groupe12.entities;
 
 public class ShopFrame {
 
-    private int id;
-    private int idShop;
+    private long id;
+    private long idShop;
     String filePath;
 
-    public ShopFrame(int id, int idShop, String filePath){
+    public ShopFrame(long id, long idShop, String filePath){
         this.id = id;
         this.idShop = idShop;
         this.filePath = filePath;
     }
 
-    public int getId() {
+    public long getId() {
         return id;
     }
 
-    public int getIdShop() {
+    public long getIdShop() {
         return idShop;
     }
 
@@ -26,11 +26,11 @@ public class ShopFrame {
 
     //TODO : Setter methods should update the DB ?
 
-    public void setId(int id) {
+    public void setId(long id) {
         this.id = id;
     }
 
-    public void setIdShop(int idShop) {
+    public void setIdShop(long idShop) {
         this.idShop = idShop;
     }
